+++ src/__tests__/mocks/notifications/transaction-canceled.mock.ts
@@ -160,6 +160,7 @@ export const TransactionCanceledMock: IEventsResponse<ITransactionNotificationRe
         subtotal: '1319900',
         grand_total: '1437041',
         currency_code: 'USD',
+        retained_fee: '2',
       },
       adjusted_payout_totals: null,
     },
@@ -225,6 +226,7 @@ export const TransactionCanceledMockExpectation = {
         subtotal: '1319900',
         tax: '117141',
         total: '1437041',
+        retainedFee: '2',
       },
       lineItems: [
         {